chore: :sparkeles: moved survicate to app component
