+++ packages/bot-web-ui/src/app/app-content.jsx
@@ -56,28 +56,6 @@ const AppContent = observer(() => {
             r.src = t + h._hjSettings.hjid + j + h._hjSettings.hjsv;
             a.appendChild(r);
         })(window, document, 'https://static.hotjar.com/c/hotjar-', '.js?sv=');
-
-        const initSurvicate = () => {
-            if (document.getElementById('dbot-survicate')) {
-                const survicate_box = document.getElementById('survicate-box') || undefined;
-                if (survicate_box) {
-                    survicate_box.style.display = 'block';
-                }
-                return;
-            }
-            const script = document.createElement('script');
-            script.id = 'dbot-survicate';
-            script.async = true;
-            script.src = 'https://survey.survicate.com/workspaces/83b651f6b3eca1ab4551d95760fe5deb/web_surveys.js';
-            document.body.appendChild(script);
-        };
-
-        initSurvicate();
-        return () => {
-            if (document.getElementById('survicate-box')) {
-                document.getElementById('survicate-box').style.display = 'none';
-            }
-        };
     }, []);
 
     React.useEffect(() => {
